requirements.txt
streamlit
pandas
openpyxl
gspread
google-auth

